+++ tests/test_management.py
@@ -12,8 +12,7 @@ from threedigrid.management.commands.kick import kick_start
 from threedigrid.management.commands.kick import export_to
 
 
-test_file_dir = os.path.join(
-    os.getcwd(), "tests/test_files")
+test_file_dir = os.path.join(os.getcwd(), "tests/test_files")
 
 # the testfile is a copy of the v2_bergermeer gridadmin file
 grid_admin_h5_file = os.path.join(test_file_dir, "gridadmin.h5")
@@ -22,15 +21,15 @@ grid_admin_h5_file = os.path.join(test_file_dir, "gridadmin.h5")
 def test_model_overview():
     # smoke test
     txt = help_texts.model_overview(grid_admin_h5_file)
-    assert txt != ''
+    assert txt != ""
 
 
 def test_command_kick_start():
     # smoke test
     runner = CliRunner()
     result = runner.invoke(
-        kick_start,
-        [u"--grid-file={}".format(grid_admin_h5_file), u"--no-ipy"])
+        kick_start, [u"--grid-file={}".format(grid_admin_h5_file), u"--no-ipy"]
+    )
     assert result.exit_code == 0
 
 
@@ -43,12 +42,13 @@ def test_command_export_to():
     runner = CliRunner()
     result = runner.invoke(
         export_to,
-        [u"--grid-file={}".format(grid_admin_h5_file),
-         u"--file-type=shape",
-         u"--output-file={}".format(test_shp),
-         u"--model=lines",
-         u"--subset=1d_all",
-         ]
+        [
+            u"--grid-file={}".format(grid_admin_h5_file),
+            u"--file-type=shape",
+            u"--output-file={}".format(test_shp),
+            u"--model=lines",
+            u"--subset=1d_all",
+        ],
     )
     assert result.exit_code == 0
-    assert (os.path.exists(test_shp))
+    assert os.path.exists(test_shp)